moved SpringProfileConstants to valid OASP package, keep deprecated original for compatibility.
