+++ samples/core/src/test/java/io/oasp/gastronomy/restaurant/general/common/base/AbstractRestServiceTest.java
@@ -16,7 +16,7 @@ import io.oasp.gastronomy.restaurant.general.common.DbTestHelper;
 import io.oasp.gastronomy.restaurant.general.common.RestTestClientBuilder;
 import io.oasp.gastronomy.restaurant.general.common.SecurityTestHelper;
 import io.oasp.gastronomy.restaurant.general.configuration.RestaurantTestConfig;
-import io.oasp.module.basic.configuration.SpringProfileConstants;
+import io.oasp.module.basic.common.api.config.SpringProfileConstants;
 import io.oasp.module.test.common.base.SubsystemTest;
 
 /**
